# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Outdoor Event Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/financials:
    get:
      operationId: getEventFinancials
      tags: [financials]
      summary: Get financial summary for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventFinancials"
        "404":
          description: Event not found
    patch:
      operationId: updateEventFinancials
      tags: [financials]
      summary: Update financial data for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventFinancialsInput"
      responses:
        "200":
          description: Updated financials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventFinancials"
  /events/{id}/financials/export:
    get:
      operationId: exportEventFinancials
      tags: [financials]
      summary: Download event financials as CSV
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "404":
          description: Event not found
...
  /participants:
    get:
      operationId: listParticipants
      tags: [participants]
      summary: List participants ordered by points
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Participant"
  /participants/{phone}/events:
    get:
      operationId: getParticipantEvents
      tags: [participants]
      summary: Get events for a participant by phone
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Participant event history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParticipantEvent"
  /participants/{phone}:
    get:
      operationId: getParticipant
      tags: [participants]
      summary: Get a participant by phone
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        "404":
          description: Not found
    patch:
      operationId: updateParticipant
      tags: [participants]
      summary: Update a participant's profile fields
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateParticipantInput"
      responses:
        "200":
          description: Updated participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        "404":
          description: Not found
  /participants/{phone}/phone:
    patch:
      operationId: changeParticipantPhone
      tags: [participants]
      summary: Change a participant's phone number (cascades to registrations)
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePhoneInput"
      responses:
        "200":
          description: Updated participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        "404":
          description: Not found
        "409":
          description: Phone already in use
...
  /media-banners:
    get:
      operationId: listMediaBanners
      tags: [media-banners]
      summary: List media banners
      parameters:
        - name: activeOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Media banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaBanner"
    post:
      operationId: createMediaBanner
      tags: [media-banners]
      summary: Create a media banner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaBannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaBanner"
  /media-banners/reorder:
    patch:
      operationId: reorderMediaBanners
      tags: [media-banners]
      summary: Reorder media banners
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/BannerReorderItem"
      responses:
        "204":
          description: Reordered
  /media-banners/{id}:
    get:
      operationId: getMediaBanner
      tags: [media-banners]
      summary: Get a single media banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaBanner"
        "404":
          description: Not found
    patch:
      operationId: updateMediaBanner
      tags: [media-banners]
      summary: Update a media banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaBannerInput"
      responses:
        "200":
          description: Updated banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaBanner"
    delete:
      operationId: deleteMediaBanner
      tags: [media-banners]
      summary: Delete a media banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ParticipantEvent:
      type: object
      required: [id, title, category, date, location, status, registrationStatus, registeredAt, pointsValue]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [cycling, hiking, summer-night, walking]
        date:
          type: string
          format: date-time
        location:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
        registrationStatus:
          type: string
          enum: [pending, confirmed, waitlist, cancelled]
        registeredAt:
          type: string
          format: date-time
        pointsValue:
          type: integer
        referralToken:
          type: ["string", "null"]
        registrationToken:
          type: string
    Participant:
      type: object
      required: [id, email, name, totalPoints, totalEvents, referralCount, joinedAt]
      properties:
        id:
          type: integer
        phone:
          type: ["string", "null"]
        email:
          type: string
        name:
          type: string
        totalPoints:
          type: integer
        totalEvents:
          type: integer
        referralCount:
          type: integer
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        waiverAcceptedAt:
          type: ["string", "null"]
          format: date-time
        joinedAt:
          type: string
          format: date-time
...
    MediaBanner:
      type: object
      required: [id, type, url, displayOrder, isActive, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [image, video]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        displayOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MediaBannerInput:
      type: object
      required: [url]
      properties:
        type:
          type: string
          enum: [image, video]
        url:
          type: string
          minLength: 1
        thumbnailUrl:
          type: string
        title:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
    BannerReorderItem:
      type: object
      required: [id, displayOrder]
      properties:
        id:
          type: integer
        displayOrder:
          type: integer
    EventFinancials:
      type: object
      required: [id, eventId, pricePerPerson, totalCollected, referralDiscounts, manualDiscounts, promoDiscounts, netRevenue, updatedAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        pricePerPerson:
          type: integer
        totalCollected:
          type: integer
        referralDiscounts:
          type: integer
        manualDiscounts:
          type: integer
        promoDiscounts:
          type: integer
        netRevenue:
          type: integer
        notes:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
    EventFinancialsInput:
      type: object
      properties:
        pricePerPerson:
          type: integer
        totalCollected:
          type: integer
        referralDiscounts:
          type: integer
        manualDiscounts:
          type: integer
        promoDiscounts:
          type: integer
        notes:
          type: string
...
    UpdateParticipantInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        totalPoints:
          type: integer
    ChangePhoneInput:
      type: object
      required: [newPhone]
      properties:
        newPhone:
          type: string
...